updtae

+++ app/Console/Commands/DeleteMessage.php
@@ -34,7 +34,7 @@ class DeleteMessage extends Command
             $telegram_services = new TelegramServices($bot_user->tokrn);
             $message = MessageAdmin::where("message_id", $message_id)->first();
             if ($message)
-                $telegram_services->deleteMessage(data_get($message,"user_id"), $message_id);
+                $telegram_services->deleteMessage((int)data_get($message,"user_id"), $message_id);
         }
 
     }
